Add keyword search
Right now if you search something like "green laptop" it'll return all results which contain one of the two. This can be fixed if you search ""green" "laptop"" instead, forcing both to be in the text search. But I don't think this will matter for general text search.

We can set a score threshold for the text search later